add: Hooks in order to further reduce launchwindow logic
LaunchWindow could still be further reduced, but honestly from 1600 lines to 650  lines is a win
