+++ src/components/launch/popovers/MicPopover.tsx
@@ -1,0 +1,94 @@
+import { MicrophoneSlash as MicOff, SpeakerHigh as Volume2, SpeakerX as VolumeX } from "@phosphor-icons/react";
+import { useScopedT } from "@/contexts/I18nContext";
+import { DropdownItem, HudPopover, MicDeviceRow } from "./PopoverScaffold";
+import { useLaunchPopoverCoordinator } from "./LaunchPopoverCoordinator";
+import type { DeviceOption } from "./types";
+import type { ReactNode } from "react";
+import styles from "../LaunchWindow.module.css";
+
+const POPOVER_ID = "mic";
+
+export function MicPopover({
+	trigger,
+	disabled,
+	systemAudioEnabled,
+	onToggleSystemAudio,
+	microphoneEnabled,
+	onDisableMicrophone,
+	devices,
+	microphoneDeviceId,
+	selectedDeviceId,
+	onSelectDevice,
+}: {
+	trigger: ReactNode;
+	disabled?: boolean;
+	systemAudioEnabled: boolean;
+	onToggleSystemAudio: () => void;
+	microphoneEnabled: boolean;
+	onDisableMicrophone: () => void;
+	devices: DeviceOption[];
+	microphoneDeviceId?: string;
+	selectedDeviceId?: string;
+	onSelectDevice: (deviceId: string) => void;
+}) {
+	const t = useScopedT("launch");
+	const { isOpen, requestOpen, requestClose } = useLaunchPopoverCoordinator();
+	const open = isOpen(POPOVER_ID);
+
+	return (
+		<HudPopover
+			open={open}
+			onOpenChange={(nextOpen) => {
+				if (!nextOpen) {
+					requestClose(POPOVER_ID);
+					return;
+				}
+				if (disabled) {
+					return;
+				}
+				requestOpen(POPOVER_ID);
+			}}
+			trigger={trigger}
+			align="start"
+		>
+			<div className={styles.ddLabel}>{t("recording.microphone")}</div>
+			<DropdownItem
+				icon={systemAudioEnabled ? <Volume2 size={16} /> : <VolumeX size={16} />}
+				selected={systemAudioEnabled}
+				onClick={onToggleSystemAudio}
+			>
+				{systemAudioEnabled
+					? t("recording.disableSystemAudio")
+					: t("recording.enableSystemAudio")}
+			</DropdownItem>
+			{microphoneEnabled && (
+				<DropdownItem
+					icon={<MicOff size={16} />}
+					onClick={() => {
+						onDisableMicrophone();
+						requestClose(POPOVER_ID);
+					}}
+				>
+					{t("recording.turnOffMicrophone")}
+				</DropdownItem>
+			)}
+			{!microphoneEnabled && (
+				<div className="px-3 py-2 text-xs text-[#6b6b78]">{t("recording.selectMicToEnable")}</div>
+			)}
+			{devices.map((device) => (
+				<MicDeviceRow
+					key={device.deviceId}
+					device={device}
+					selected={
+						microphoneEnabled &&
+						(microphoneDeviceId === device.deviceId || selectedDeviceId === device.deviceId)
+					}
+					onSelect={() => onSelectDevice(device.deviceId)}
+				/>
+			))}
+			{devices.length === 0 && (
+				<div className="text-center text-xs text-[#6b6b78] py-4">{t("recording.noMicrophonesFound")}</div>
+			)}
+		</HudPopover>
+	);
+}
